requirements.txt
Pillow>=11.3.0
ttkbootstrap>=1.10.1
psutil
pydantic>=2
pyyaml
requests
